// src/types/user.ts
export type UserForm = {
  address: string;
  personalId: string;
  freelanceWorkId: string;
  bankNumber: string;
  bankName: string;
  vatCode: string | null;
};

export type UserProfile = {
  userId: string;
  firstName: string;
  lastName: string;
  address: string;
  personalId: string;
  freelanceWorkId: string;
  bankNumber: string;
  bankName: string;
  vatCode: string | null;
};